Renamed DevSkim project and associated dependencies to remove potential name confusion with the new DevSkim redesign away from .net so this fork as it were becomes entirely part of ApplicationInspector only.
