+++ AppInspector/html/partials/_report_summary.liquid
@@ -122,7 +122,7 @@
                             </td>
                             <td>
                               <a href="#" onclick="SortbySeverity();">
-                                <strong>Severity</strong>
+                                <strong>Importance</strong>
                               </a>
                             </td>
                           </tr>
@@ -157,7 +157,7 @@
                             </td>
                             <td>
                               <a href="#" onclick="SortbySeverity();">
-                                <strong>Severity</strong>
+                                <strong>Importance</strong>
                               </a>
                             </td>
                           </tr>
@@ -191,7 +191,7 @@
                             </td>
                             <td>
                               <a href="#" onclick="SortbySeverity();">
-                                <strong>Severity</strong>
+                                <strong>Importance</strong>
                               </a>
                             </td>
                           </tr>
