+++ RulesEngine/Confidence.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.DevSkim
+namespace RulesEngine
 {
     [Flags]
     public enum Confidence { Low = 1, Medium = 2, High = 4 }
