+++ RulesEngine/Location.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Microsoft.DevSkim
+namespace RulesEngine
 {
     public class Location
     {
